retrieval/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
etelemetry==0.3.1
faiss-cpu==1.8.0
fastapi==0.110.3
filelock==3.14.0
fsspec==2024.3.1
httplib2==0.22.0
huggingface-hub==0.22.2
idna==3.7
isodate==0.6.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.48
langchain-text-splitters==0.0.1
langsmith==0.1.52
looseversion==1.3.0
lxml==5.2.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
orjson==3.10.2
packaging==23.2
pandas==2.2.2
pathlib==1.0.1
prov==2.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydot==2.0.0
pyparsing==3.1.2
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
scipy==1.13.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
traits==6.3.2
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
